\name{RcppColors-package}
\alias{RcppColors-package}
\alias{RcppColors}
\docType{package}
\title{
  'C++' header files for conversion from some color spaces to 
  hexadecimal.
}
\description{
  This package is mainly intended to be used with 'Rcpp'.
}
\details{
  See README for a description of the available 'C++' functions and 
  how to use the package.
}
\author{
Stéphane Laurent.

Maintainer: Stéphane Laurent <laurent_step@outlook.fr>
}
\keyword{ package }
