% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsluv.R
\name{hsluv}
\alias{hsluv}
\alias{hsluv2rgb}
\title{HSLuv color specification}
\usage{
hsluv(h = 360, s = 100, l = 100)

hsluv2rgb(h = 360, s = 100, l = 100)
}
\arguments{
\item{h}{the hue, a number between \code{0} and \code{360}}

\item{s}{the saturation, a number between \code{0} and \code{100}}

\item{l}{the lightness, a number between \code{0} and \code{100}}
}
\value{
The \code{hsluv} function returns a hexadecimal string representing 
  a color, and the \code{hsluv2rgb} returns the RGB coordinates of this 
  color, a named vector of three integers between \code{0} and \code{255}.
}
\description{
Converts a color given in HSLuv coordinates to a hexadecimal 
  string or a RGB color specification
}
\examples{
xx
}
